+++ ai.hh
@@ -1,9 +1,28 @@
 #pragma once
 
+/// \file ai.hh
+/// \author Egor Starobinskii
+/// \date 11 Aug 2018
+/// \brief Single-header C++ Library
+/// \details This is a single-header C++ Library from Ailurus Studio that 
+/// brings you extra time to admire life instead of coding the same functions 
+/// again and again.
+/// \see https://starobinskii.github.io/AiLibrary
+/// \see https://github.com/starobinskii/AiLibrary
+/// \see https://snapcraft.io/ailibrary
+/// \see https://ailurus.ru/
+
+/// \brief If defined, funstions will be marked as inline. 
+/// \details If defined, funstions will be marked as inline. Delete the line, 
+/// if you want to omit this behavior.
+#define INLINE inline
+
 #include <array>
 #include <cmath>
 #include <chrono>
+#include <limits>
 #include <memory>
+#include <random>
 #include <string>
 #include <vector>
 #include <cstdlib>
@@ -21,18 +40,44 @@
     #include <dirent.h>
 #endif
 
+/// \namespace ai
+/// \brief Main namespace
+/// \details Main namespace containing all the functions of AiLibrary
+/// \defgroup AiNamespace AiLibrary Namespace
+/// \brief Main group
+/// \details Main group containing all the functions of AiLibrary
+/// \{
 namespace ai{
-    /*/ **************************************************************** /*/
+    /// \defgroup LibraryInfo Library Info
+    /// \brief Functions providing info about the library
+    /// \details Group of functions that provides information about this 
+    /// library
+    /// \{
 
-    inline std::string getVersion(){
-        return "1.1.0";
+    /// \fn std::string getVersion();
+    /// \brief Gets version of the lib
+    /// \details This function returns version of the AiLibrary (we use SemVer)
+    /// \return Version as a string
+    INLINE std::string getVersion(){
+        return "1.2.0";
     }
 
+    /// \} End of LibraryInfo Group
 
-    /*/ **************************************************************** /*/
+    /// \defgroup StringFunctions String Functions
+    /// \brief Functions providing additional methods to work with strings
+    /// \details Group of functions that allows you to freely manipulate and 
+    /// analyze strings
+    /// \{
 
+    /// \fn std::string string(const T value);
+    /// \brief Converts input into a string
+    /// \details This function converts your variable into std::string using 
+    /// std::ostringstream
+    /// \param value Your variable to convert
+    /// \return std::string copy of your input
     template<typename T>
-    std::string string(const T value){
+    INLINE std::string string(const T value){
         std::ostringstream stream;
 
         stream << value;
@@ -40,49 +85,199 @@ namespace ai{
         return stream.str();
     }
 
-    /*/ **************************************************************** /*/
+    /// \fn bool hasPrefix(const std::string &text, const std::string &prefix);
+    /// \brief Checks if a string begins with a substring
+    /// \details This function checks if a string begins with a substring
+    /// \param text Your string to test
+    /// \param prefix Your substring
+    /// \return True if string begins with a substring, false otherwise
+    INLINE bool hasPrefix(const std::string &text, const std::string &prefix){
+        return text.size() >= prefix.size()
+            && 0 == text.compare(0, prefix.size(), prefix);
+    }
 
-    inline std::size_t counter(std::size_t value = 0){
+    /// \fn bool hasSuffix(const std::string &text, const std::string &suffix);
+    /// \brief Checks if a string ends with a substring
+    /// \details This function checks if a string ends with a substring
+    /// \param text Your string to test
+    /// \param suffix Your substring
+    /// \return True if string ends with a substring, false otherwise
+    INLINE bool hasSuffix(const std::string &text, const std::string &suffix){
+        return text.size() >= suffix.size() &&
+            text.compare(
+                text.size() - suffix.size(), 
+                suffix.size(), 
+                suffix
+            ) == 0;
+    }
+
+    /// \fn bool contains(const std::string &text, const std::string 
+    /// &substring);
+    /// \brief Checks if a string contains a substring
+    /// \details This function checks if a string contains a substring
+    /// \param text Your string to test
+    /// \param substring Your substring
+    /// \return True if string contains a substring, false otherwise
+    INLINE bool contains(
+        const std::string &text,
+        const std::string &substring
+    ){
+        return std::string::npos != text.find(substring);
+    }
+
+    /// \fn std::string replace(std::string text, const std::string &substring,
+    /// const std::string &replacement)
+    /// \brief Replaces all occurrences of a substring in a copy of the initial 
+    /// string with your text
+    /// \details This function replaces all occurrences of a substring in a 
+    /// string with your text and return the result. Initial string stays the 
+    /// same
+    /// \param text Your string to modify
+    /// \param substring Your substring to find in the string
+    /// \param replacement Replacement for all the substring occurrences
+    /// \return Modified copy of the initial string
+    INLINE std::string replace(
+        std::string text,
+        const std::string &substring,
+        const std::string &replacement
+    ){
+        std::size_t position = text.find(substring);
+        std::size_t substringSize = substring.size();
+        std::size_t replacementSize = replacement.size();
+        
+        while(std::string::npos != position){
+            text.replace(position, substringSize, replacement);
+            position = text.find(substring, position + replacementSize);
+        }
+        return text;
+    }
+
+    /// \fn void applyReplace(std::string &text, const std::string 
+    /// &substring, const std::string &replacement)
+    /// \brief Modifies your string by replacing all occurrences of a substring
+    /// string with your text
+    /// \details This function replaces all occurrences of a substring in a 
+    /// string with your text (modifies the initial string)
+    /// \param text Your string to modify
+    /// \param substring Your substring to find in the string
+    /// \param replacement Replacement for all the substring occurrences
+    INLINE void applyReplace(
+        std::string &text,
+        const std::string &substring,
+        const std::string &replacement
+    ){
+        std::size_t position = text.find(substring);
+        std::size_t substringSize = substring.size();
+        std::size_t replacementSize = replacement.size();
+        
+        while(std::string::npos != position){
+            text.replace(position, substringSize, replacement);
+            position = text.find(substring, position + replacementSize);
+        }
+    }
+
+    /// \fn bool equal(const char *charString, const std::string string1);
+    /// \brief Checks if a char string is equal to a std::string
+    /// \details This function compares a char string with a std::string
+    /// \param charString Your char string to compare
+    /// \param string1 Your std::string to compare
+    /// \return True if strings are equal, false otherwise
+    INLINE bool equal(const char *charString, const std::string string1){
+        const std::string string2(charString);
+        
+        return string1 == string2;
+    }
+
+    /// \} End of StringFunctions Group
+
+    /// \defgroup DebugFunctions Debug Functions
+    /// \brief Functions providing some help to debug programs
+    /// \details Group of functions that can be useful if you want to debug 
+    /// your code
+    /// \{
+
+    /// \fn std::size_t counter(std::size_t value = 0);
+    /// \brief Returns ID starting from zero or the specified value
+    /// \details This function returns ID (increases it at each call) starting 
+    /// from zero or the specified non-negative value
+    /// \param value The non-negative  value to which the counter should be 
+    /// reset (optional)
+    /// \return Counter value
+    INLINE std::size_t counter(std::size_t value = 0){
         static std::size_t count = 0;
+
+        ++count;
 
         if(0 != value){
             count = value;
         }
-
-        ++count;
 
         return count;
     }
 
-    inline std::string marker(std::size_t value = 0){
+    /// \fn std::string marker(std::size_t value = 0);
+    /// \brief Returns a string containing the word "Marker" and its ID
+    /// \details This function returns a string containing the word "marker" 
+    /// and its ID (increases it at each call). ID specified in the same way 
+    /// as in the function counter()
+    /// \param value The non-negative  value to which the counter should be 
+    /// reset (optional)
+    /// \return Marker string
+    INLINE std::string marker(std::size_t value = 0){
         static std::size_t count = 0;
+
+        ++count;
 
         if(0 != value){
             count = value;
         }
 
-        ++count;
-
         return "Marker #" + ai::string(count);
     }
 
-    inline void printMarker(std::size_t value = 0){
+    /// \fn void printMarker(std::size_t value = 0);
+    /// \brief Calls marker() and prints result to stdout
+    /// \details This function calls marker() and prints result to stdout
+    /// \param value The non-negative  value to which the counter should be 
+    /// reset (optional)
+    INLINE void printMarker(std::size_t value = 0){
         std::cout << marker(value) << std::endl;
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of DebugFunctions Group
 
+    /// \defgroup MathFunctions Math Functions
+    /// \brief Functions providing different math methods and checks
+    /// \details Group of functions that completes standard C++ mathematical 
+    /// methods and helps with matrices and vectors
+    /// \{
+
+    /// \fn T sign(const T value);
+    /// \brief Returns signum of the value
+    /// \details This function returns signum of the value (usign copysign())
+    /// \param value The number to which signum is applied
+    /// \return -1 for negative values, +1 for positive, 0 for zero
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T sign(const T a){
-        if(0 == a){
+    INLINE T sign(const T value){
+        if(0 == value){
             return (T) 0;
         }
 
-        return copysign(1, a);
+        return copysign(1, value);
     }
 
+    /// \fn T min(const T a, const T b);
+    /// \brief Returns minimum of two values
+    /// \details This function compares two values and returns a minimum
+    /// \param a First number
+    /// \param b Second number
+    /// \return Minimum of two values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T min(const T a, const T b){
+    INLINE T min(const T a, const T b){
         if(a > b){
             return b;
         }
@@ -90,8 +285,16 @@ namespace ai{
         return a;
     }
 
+    /// \fn T max(const T a, const T b);
+    /// \brief Returns maximum of two values
+    /// \details This function compares two values and returns a maximum
+    /// \param a First number
+    /// \param b Second number
+    /// \return Maximum of two values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T max(const T a, const T b){
+    INLINE T max(const T a, const T b){
         if(a < b){
             return b;
         }
@@ -99,8 +302,16 @@ namespace ai{
         return a;
     }
 
+    /// \fn T min(const std::vector<T> &input);
+    /// \brief Returns minimum of vector values
+    /// \details This function compares vector values using min() and returns 
+    /// a minimum
+    /// \param input Vector to search for a minimum value
+    /// \return Minimum of vector values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T min(const std::vector<T> input){
+    INLINE T min(const std::vector<T> &input){
         T minimum = input[0];
         
         for(std::size_t i = 1; i < input.size(); ++i){
@@ -110,8 +321,16 @@ namespace ai{
         return minimum;
     }
 
+    /// \fn T max(const std::vector<T> &input);
+    /// \brief Returns maximum of vector values
+    /// \details This function compares vector values using max() and returns 
+    /// a maximum
+    /// \param input Vector to search for a maximum value
+    /// \return Maximum of vector values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T max(const std::vector<T> input){
+    INLINE T max(const std::vector<T> &input){
         T maximum = input[0];
         
         for(std::size_t i = 1; i < input.size(); ++i){
@@ -121,8 +340,16 @@ namespace ai{
         return maximum;
     }
 
+    /// \fn T min(const std::vector< std::vector<T> > &input);
+    /// \brief Returns minimum of matrix values
+    /// \details This function compares matrix values using min() and returns 
+    /// a minimum
+    /// \param input Matrix to search for a minimum value
+    /// \return Minimum of matrix values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T min(const std::vector< std::vector<T> > input){
+    INLINE T min(const std::vector< std::vector<T> > &input){
         T minimum = input[0][0];
 
         for(std::size_t i = 0; i < input.size(); ++i){
@@ -134,8 +361,16 @@ namespace ai{
         return minimum;
     }
 
+    /// \fn T max(const std::vector< std::vector<T> > &input);
+    /// \brief Returns maximum of matrix values
+    /// \details This function compares matrix values using max() and returns 
+    /// a maximum
+    /// \param input Matrix to search for a maximum value
+    /// \return Maximum of matrix values
+    /// \tparam T A number type
+    /// \todo Add tests
     template<typename T>
-    T max(const std::vector< std::vector<T> > input){
+    INLINE T max(const std::vector< std::vector<T> > &input){
         T maximum = input[0][0];
 
         for(std::size_t i = 0; i < input.size(); ++i){
@@ -147,13 +382,178 @@ namespace ai{
         return maximum;
     }
 
-    /*/ **************************************************************** /*/
+    /// \fn bool isSquare(const T value);
+    /// \brief Checks if number is square
+    /// \details This function checks if number is square
+    /// \param value Number to test
+    /// \return True if number is a square, false otherwise
+    /// \tparam T A number type
+    /// \todo Add tests
+    template<typename T>
+    INLINE bool isSquare(const T value){
+        if(0 >= value){
+            return false;
+        }
 
-    inline std::chrono::high_resolution_clock::time_point time(){
+        T root = (T) round(sqrt(value));
+
+        return value == root * root;
+    }
+
+    /// \fn bool isSquare(const std::vector< std::vector<T> > matrix);
+    /// \brief Checks if matrix is square
+    /// \details This function checks if matrix is square
+    /// \param matrix Matrix to test
+    /// \return True if matrix is square, false otherwise
+    /// \todo Add tests
+    template<typename T>
+    INLINE bool isSquare(const std::vector< std::vector<T> > matrix){
+        return 0 < matrix.size() && matrix.size() == matrix[0].size();
+    }
+
+    /// \todo Add description. Add tests. Check for real or int
+    template<typename T>
+    INLINE void generateRandomVector(
+        std::vector<T> &vector,
+        const std::size_t length,
+        const T min = std::numeric_limits<T>::min(),
+        const T max = std::numeric_limits<T>::max()
+    ){
+        std::uniform_real_distribution<T> distribution(min, max);
+        std::random_device device;
+        srand(device() * time(0));
+
+        vector.resize(length);
+
+        std::generate(
+            vector.begin(),
+            vector.end(),
+            [&device, &distribution](){
+                return distribution(device);
+            }
+        );
+    }
+
+    /// \todo Add description. Add tests
+    template<typename T>
+    INLINE void generateRandomMatrix(
+        std::vector< std::vector<T> > &matrix,
+        const std::size_t xSize,
+        const std::size_t ySize,
+        const T min = std::numeric_limits<T>::min(),
+        const T max = std::numeric_limits<T>::max()
+    ){
+        matrix.resize(xSize);
+
+        for(size_t i = 0; i < xSize; ++i){
+            matrix[i].resize(ySize);
+
+            std::vector<T> row;
+
+            ai::generateRandomVector(row, ySize, min, max);
+
+            matrix[i] = row;
+        }
+    }
+
+    /// \todo Add description. Add tests
+    template<typename T>
+    INLINE void generateRandomMatrix(
+        std::vector<T> &vector,
+        const std::size_t size,
+        const T min = std::numeric_limits<T>::min(),
+        const T max = std::numeric_limits<T>::max()
+    ){
+        generateRandomMatrix(vector, size, size, min, max);
+    }
+
+    /// \fn void translateMatrixIntoVector( std::vector< std::vector<T> > 
+    /// &matrix, std::vector<T> &vector);
+    /// \brief Elongates matrix into a vector
+    /// \details This function converts the matrix into a vector, writing each 
+    /// row one after another in a line
+    /// \param matrix Matrix to tranform
+    /// \param vector Vector to store the result
+    /// \todo Add tests
+    template<typename T>
+    INLINE void translateMatrixIntoVector(
+        std::vector< std::vector<T> > &matrix,
+        std::vector<T> &vector
+    ){
+        for(std::size_t i = 0; i < matrix.size(); ++i){
+            for(std::size_t j = 0; j < matrix[i].size(); ++j){
+                vector.push_back(matrix[i][j]);
+            }
+        }
+    }
+
+    /// \fn void translateVectorIntoSquareMatrix(std::vector<T> &vector, 
+    /// std::vector< std::vector<T> > &matrix);
+    /// \brief Transform vector in a square matrix (if possible)
+    /// \details This function converts the vector into a matrix, if possible. 
+    /// Otherwise, an exception will be thrown at runtime
+    /// \param vector Vector to tranform
+    /// \param matrix Matrix to store the result
+    /// \exception std::runtime_error If \p matrix is not square
+    /// \todo Add tests
+    template<typename T>
+    INLINE void translateVectorIntoSquareMatrix(
+        std::vector<T> &vector,
+        std::vector< std::vector<T> > &matrix
+    ){
+        if(!isSquare(vector.size())){
+            throw std::runtime_error(
+                ai::string("Exception while converting vector into the square ")
+                + ai::string("matrix: vector size should be square") 
+            );
+        }
+
+        std::size_t size = (std::size_t) round(sqrt(vector.size()));
+
+        for(std::size_t i = 0; i < size; ++i){
+            std::vector<T> row;
+
+            for(std::size_t j = 0; j < size; ++j){
+                row.push_back(vector[i * size + j]);
+            }
+
+            matrix.push_back(row);
+        }
+    }
+
+    /// \} End of MathFunctions Group
+
+    /// \defgroup TimeFunctions Time Functions
+    /// \brief Functions providing methods to measure time
+    /// \details Group of functions that is in help when you need to measure 
+    /// execution time of some code section
+    /// \{
+
+    /// \fn std::chrono::high_resolution_clock::time_point time();
+    /// \brief Returns current time point
+    /// \details This function returns current time point using std::chrono
+    /// \return std::chrono::high_resolution_clock entity
+    /// \todo Add tests
+    INLINE std::chrono::high_resolution_clock::time_point time(){
         return std::chrono::high_resolution_clock::now();
     }
 
-    inline double duration(
+    /// \fn double duration(const 
+    /// std::chrono::high_resolution_clock::time_point start, const 
+    /// std::chrono::high_resolution_clock::time_point finish, const 
+    /// std::string scale = std::string("ms"));
+    /// \brief Returns the difference between two time points
+    /// \details This function returns the difference between two time points 
+    /// in seconds, milliseconds or microseconds using std::chrono (handy to
+    /// measure functions and code blocks)
+    /// \param start Time point
+    /// \param finish Time point
+    /// \param scale Optional. If \p scale equals to "s", function will return 
+    /// difference in seconds. If \p scale equals to "us" – in microseconds. 
+    /// Otherwise – in microseconds.
+    /// \return Difference between points
+    /// \todo Add tests
+    INLINE double duration(
         const std::chrono::high_resolution_clock::time_point start,
         const std::chrono::high_resolution_clock::time_point finish,
         const std::string scale = std::string("ms")
@@ -172,44 +572,16 @@ namespace ai{
             <std::chrono::milliseconds> (finish - start).count();
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of TimeFunctions Group
 
-    inline bool hasPrefix(const std::string &str, const std::string &prefix){
-        return str.size() >= prefix.size()
-            && 0 == str.compare(0, prefix.size(), prefix);
-    }
+    /// \defgroup ParameterFunctions Parameter Functions
+    /// \brief Functions providing methods to parse and assign parameters
+    /// \details Group of functions that is useful to parse parameters (e.g., 
+    /// from command line) and assign its values to your variables
+    /// \{
 
-    inline bool hasSuffix(const std::string &str, const std::string &suffix){
-        return str.size() >= suffix.size() &&
-            str.compare(str.size() - suffix.size(), suffix.size(), suffix) == 0;
-    }
-
-   inline bool contains(
-       std::string text,
-       const std::string &substring
-   ){
-       return std::string::npos != text.find(substring);
-   }
-
-    inline std::string replace(
-        std::string text,
-        const std::string &substring,
-        const std::string &replacement
-    ){
-        std::size_t position = text.find(substring);
-        std::size_t substringSize = substring.size();
-        std::size_t replacementSize = replacement.size();
-        
-        while(std::string::npos != position){
-            text.replace(position, substringSize, replacement);
-            position = text.find(substring, position + replacementSize);
-        }
-        return text;
-    }
-
-    /*/ **************************************************************** /*/
-
-    inline std::string parseParameter(
+    /// \todo Add description. Add tests
+    INLINE std::string parseParameter(
         const char *input,
         const std::string name
     ){
@@ -223,8 +595,9 @@ namespace ai{
         return NULL;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void assignFromVectorByIntervalCondition(
+    INLINE void assignFromVectorByIntervalCondition(
         T &value,
         const T parameter,
         const std::vector< std::vector<T> > intervals
@@ -238,8 +611,9 @@ namespace ai{
         }
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void assignFromVectorByIntervalCondition(
+    INLINE void assignFromVectorByIntervalCondition(
         T &firstValue,
         T &secondValue,
         const T parameter,
@@ -255,7 +629,8 @@ namespace ai{
         }
     }
 
-    inline bool assignBooleanParameter(
+    /// \todo Add description. Add tests
+    INLINE bool assignBooleanParameter(
         const char *input,
         const std::string name,
         bool &value
@@ -269,7 +644,8 @@ namespace ai{
        return false;
    }
 
-    inline bool assignCharParameter(
+    /// \todo Add description. Add tests
+    INLINE bool assignCharParameter(
         const char *input,
         const std::string name,
         char &value
@@ -291,7 +667,8 @@ namespace ai{
         return false;
     }
 
-    inline bool assignStringParameter(
+    /// \todo Add description. Add tests
+    INLINE bool assignStringParameter(
         const char *input,
         const std::string name,
         std::string &value
@@ -307,8 +684,9 @@ namespace ai{
         return false;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline bool assignParameter(
+    INLINE bool assignParameter(
         const char *input,
         const std::string name,
         T &value
@@ -326,7 +704,8 @@ namespace ai{
         return false;
     }
 
-    inline bool assignAbsDoubleParameter(
+    /// \todo Add description. Add tests
+    INLINE bool assignAbsDoubleParameter(
         const char *input,
         const std::string name,
         double &value
@@ -343,9 +722,40 @@ namespace ai{
         return false;
     }
 
-    /*/ **************************************************************** /*/
+    /// \todo Add description. Add tests
+    template<typename T>
+    INLINE bool assignByCheckingParameter(
+        const char *input,
+        const std::string parameter,
+        T &value,
+        const T supposed
+    ){
+        if(ai::equal(input, parameter)){
+            value = supposed;
+            
+            return true;
+        }
+        
+        return false;
+    }
 
-    inline void showProgressBar(
+    /// \} End of ParameterFunctions Group
+
+    /// \defgroup InterfaceFunctions Interface Functions
+    /// \brief Functions providing some style to your console output
+    /// \details Group of functions that animate your console with a style-ish
+    /// interface
+    /// \{
+
+    /// \fn void showProgressBar(double progress, const std::size_t 
+    /// screenWidth = 80);
+    /// \brief Prints a simple progress-bar to stdout
+    /// \details This function prints a simple progress-bar to stdout
+    /// \param progress Number from 0 to 1 to 
+    /// \param screenWidth Optional. The maximum length of the progress-bar in 
+    /// the symbols. Default is 80, should be at least 20
+    /// \exception std::runtime_error If \p screenWidth is less than 20
+    INLINE void showProgressBar(
         double progress,
         const std::size_t screenWidth = 80
     ){
@@ -378,17 +788,42 @@ namespace ai{
             << " " << progress * 100. << "%";
     }
 
-    inline void printLine(
+    /// \fn void printLine(const std::string line, const std::size_t 
+    /// screenWidth = 80);
+    /// \brief Prints a line with specified width
+    /// \details This function prints a line with specified width. Can be 
+    /// useful to print output when progress-bar is displayed (function 
+    /// showProgressBar()).
+    /// \param line Text to print
+    /// \param screenWidth Optional. The maximum length of the progress-bar in 
+    /// the symbols. Default is 80, should be at least 20
+    /// \exception std::runtime_error If \p screenWidth is less than 20
+    INLINE void printLine(
         const std::string line,
         const std::size_t screenWidth = 80
     ){
+        if(20 > screenWidth){
+            throw std::runtime_error(
+                ai::string("Screen width should be at least 20 (not ")
+                + ai::string(screenWidth)
+                + ai::string(")")
+            );
+        }
+        
         std::cout << "\r" << std::setw(screenWidth) << std::left << line << std::endl;
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of InterfaceFunctions Group
 
+    /// \defgroup ReadFunctions Read Functions
+    /// \brief Functions providing the ability to read your data
+    /// \details Group of functions that reads files and writes its content 
+    /// into your variables
+    /// \{
+
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void parseFileInMatrix(
+    INLINE void parseFileInMatrix(
         const std::string filename,
         const char separator,
         std::vector< std::vector<T> > &matrix
@@ -425,8 +860,9 @@ namespace ai{
         }
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void parseFileInVector(
+    INLINE void parseFileInVector(
         const std::string filename,
         const char separator,
         std::vector<T> &vector
@@ -469,7 +905,8 @@ namespace ai{
         }
     }
 
-    inline void parseFileIntoString(
+    /// \todo Add description. Add tests
+    INLINE void parseFileIntoString(
         const std::string filename,
         std::string &content
     ){
@@ -487,8 +924,9 @@ namespace ai{
         content = stream.str();
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    void accumulateFileInMatrix(
+    INLINE void accumulateFileInMatrix(
         const std::string filename,
         const char separator,
         std::vector< std::vector<T> > &matrix
@@ -508,8 +946,9 @@ namespace ai{
         }
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    void accumulateFileInVector(
+    INLINE void accumulateFileInVector(
         const std::string filename,
         const char separator,
         std::vector<T> &vector,
@@ -528,10 +967,17 @@ namespace ai{
         );
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of ReadFunctions Group
 
+    /// \defgroup PrintFunctions Print Functions
+    /// \brief Functions providing the ability to print your data
+    /// \details Group of functions that prints your data and its parameters 
+    /// for quick analysis
+    /// \{
+
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void printMatrix(
+    INLINE void printMatrix(
         const std::vector<std::vector <T> > matrix,
         const int precision = 5
     ){
@@ -561,8 +1007,9 @@ namespace ai{
             << "x" << matrix[0].size() << "]" << std::endl;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void printVector(
+    INLINE void printVector(
         const std::vector<T> vector,
         const int precision = 5
     ){
@@ -581,10 +1028,17 @@ namespace ai{
         std::cout << "}[" << vector.size() << "]" << std::endl;
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of PrintFunctions Group
 
+    /// \defgroup SaveFunctions Save Functions
+    /// \brief Functions providing the ability to save your data
+    /// \details Group of functions that will make saving data much easier 
+    /// by supporting various formats and styles
+    /// \{
+
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void saveMatrix(
+    INLINE void saveMatrix(
         const std::string filename,
         const std::vector<std::vector <T> > matrix,
         std::string comment = std::string(),
@@ -643,8 +1097,9 @@ namespace ai{
         output << suffix;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void saveVector(
+    INLINE void saveVector(
         const std::string filename,
         const std::vector <T> vector,
         std::string comment = std::string(),
@@ -696,8 +1151,9 @@ namespace ai{
         output << vector[lastIndex] << suffix;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline void saveLine(
+    INLINE void saveLine(
         const std::string filename,
         const std::string line,
         std::string comment = std::string()
@@ -718,8 +1174,9 @@ namespace ai{
         output << line;
     }
 
+    /// \todo Add description. Add tests
     template<typename T>
-    inline bool saveA3R(
+    INLINE bool saveA3R(
         const std::string filename,
         std::vector< std::vector<T> > positions,
         const double radius = 1
@@ -794,15 +1251,22 @@ namespace ai{
         return true;
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of SaveFunctions Group
 
-    inline bool folderExists(const std::string name){
+    /// \defgroup FileFunctions File Functions
+    /// \brief Functions providing the ability to check files and folders
+    /// \details Group of functions that checks and analyzes folders and files
+    /// \{
+
+    /// \todo Add description. Add tests
+    INLINE bool folderExists(const std::string name){
         struct stat buffer;
 
         return 0 == stat(name.c_str(), &buffer);
     }
 
-    inline std::size_t countLinesInFile(
+    /// \todo Add description. Add tests
+    INLINE std::size_t countLinesInFile(
         const std::string filename,
         const std::string token = std::string()
     ){
@@ -832,7 +1296,8 @@ namespace ai{
         return count;
     }
 
-    inline std::vector<std::string> listFilesWithExtension(
+    /// \todo Add description. Add tests
+    INLINE std::vector<std::string> listFilesWithExtension(
         const std::string path,
         const std::string extension,
         const std::string prefix = std::string()
@@ -858,10 +1323,21 @@ namespace ai{
         return files;
     }
 
-    /*/ **************************************************************** /*/
+    /// \} End of FileFunctions Group
 
-    inline std::string execute(const std::string command){
-        std::array<char, 128> buffer;
+    /// \defgroup ShellFunctions Shell Functions
+    /// \brief Functions providing shell support
+    /// \details Group of functions that will help you to interact with shell
+    /// and other programs
+    /// \{
+
+    /// \todo Add description. Add tests. Add more options to bufferSize
+    INLINE std::string execute(
+        const std::string command
+    ){
+        const std::size_t bufferSize = 128;
+
+        std::array<char, bufferSize> buffer;
 
         std::string result;
 
@@ -875,11 +1351,79 @@ namespace ai{
         }
 
         while(!feof(pipe.get())){
-            if(nullptr != fgets(buffer.data(), 128, pipe.get())){
+            if(nullptr != fgets(buffer.data(), bufferSize, pipe.get())){
                 result += buffer.data();
             }
         }
 
         return result;
     }
+
+    /// \} End of Shell Functions Group
 }
+/// \} End of Main Group
+
+#if defined AI_FUTURE
+    /// \brief Functions to be added in future releases
+    namespace ai{
+        INLINE bool createFolder(const std::string name);
+
+        template<typename T>
+        INLINE bool readA3R(
+            const std::string filename,
+            std::vector< std::vector<T> > &positions,
+            double &radius
+        );
+
+        template<typename T>
+        INLINE bool readA3R(
+            const std::string filename,
+            std::vector< std::vector<T> > &positions
+        );
+
+        template<typename T>
+        INLINE bool readXYZ(
+            const std::string filename,
+            std::vector< std::vector<T> > &positions
+        );
+
+        template<typename T>
+        INLINE bool saveXYZ(
+            const std::string filename,
+            std::vector< std::vector<T> > &positions
+        );
+
+        INLINE void saveLog(
+            const std::string filename,
+            std::string &log,
+            const bool timestamp = false
+        );
+
+        INLINE void saveLog(
+            const std::string filename,
+            std::vector<std::string> &logs,
+            const bool timestamp = false
+        );
+
+        template<typename T>
+        INLINE void multiplicate(
+            std::vector< std::vector<T> > &left,
+            std::vector< std::vector<T> > &right,
+            std::vector< std::vector<T> > &result,
+        );
+
+        template<typename T>
+        INLINE void multiplicate(
+            std::vector< std::vector<T> > &left,
+            std::vector<T> &right,
+            std::vector<T> &result,
+        );
+
+        template<typename T>
+        INLINE void multiplicate(
+            std::vector<T> &left,
+            std::vector<T> &right,
+            std::vector<T> &result,
+        );
+    }
+#endif