+++ docs/d5/dd6/group___library_info.js
@@ -1,0 +1,4 @@
+var group___library_info =
+[
+    [ "getVersion", "d5/dd6/group___library_info.html#gae81b57260fe20a158b0226ca9b51610e", null ]
+];