+++ docs/df/d73/group___read_functions.js
@@ -1,0 +1,8 @@
+var group___read_functions =
+[
+    [ "accumulateFileInMatrix", "df/d73/group___read_functions.html#ga9ad08eb45ba0cb425d25e8644bc33a69", null ],
+    [ "accumulateFileInVector", "df/d73/group___read_functions.html#ga6dc3993c11c7e8af3c3847b1efbe7a4b", null ],
+    [ "parseFileInMatrix", "df/d73/group___read_functions.html#gae9d45625612ddcd2581e26843f141213", null ],
+    [ "parseFileIntoString", "df/d73/group___read_functions.html#gac8b3d3c1db1973ae2d57265e435d2066", null ],
+    [ "parseFileInVector", "df/d73/group___read_functions.html#ga90ae1cfa3d587cdd7d5cb1db70f2621f", null ]
+];