+++ docs/search/all_2.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['debug_20functions',['Debug Functions',['../d3/ddf/group___debug_functions.html',1,'']]],
+  ['duration',['duration',['../dd/dd0/group___time_functions.html#ga912fcdb43dab89156260759a2fc128d6',1,'ai']]]
+];