+++ docs/search/all_7.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['inline',['INLINE',['../de/d9e/ai_8hh.html#a2eb6f9e0395b47b8d5e3eeae4fe0c116',1,'ai.hh']]],
+  ['interface_20functions',['Interface Functions',['../d2/dbc/group___interface_functions.html',1,'']]],
+  ['issquare',['isSquare',['../d5/d12/group___math_functions.html#gac521d17c2aff565fbc375f2f9c611dcd',1,'ai::isSquare(const T value)'],['../d5/d12/group___math_functions.html#gaba5e6609f9c8fd769ce6d66d46159b2b',1,'ai::isSquare(const std::vector&lt; std::vector&lt; T &gt; &gt; matrix)']]]
+];