+++ docs/search/all_8.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['library_20info',['Library Info',['../d5/dd6/group___library_info.html',1,'']]],
+  ['license_2emd',['LICENSE.md',['../dd/d79/_l_i_c_e_n_s_e_8md.html',1,'']]],
+  ['listfileswithextension',['listFilesWithExtension',['../df/dae/group___file_functions.html#ga46c6f8665ce9034fa718b3958dcfe123',1,'ai']]],
+  ['license',['LICENSE',['../d2/d35/md__l_i_c_e_n_s_e.html',1,'']]]
+];