+++ docs/search/defines_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['inline',['INLINE',['../de/d9e/ai_8hh.html#a2eb6f9e0395b47b8d5e3eeae4fe0c116',1,'ai.hh']]]
+];