+++ docs/search/functions_2.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['duration',['duration',['../dd/dd0/group___time_functions.html#ga912fcdb43dab89156260759a2fc128d6',1,'ai']]]
+];