+++ docs/search/functions_b.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['replace',['replace',['../d4/d65/group___string_functions.html#gaa3ee3ad75450b355f177173739343c01',1,'ai']]]
+];