+++ docs/search/groups_3.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['interface_20functions',['Interface Functions',['../d2/dbc/group___interface_functions.html',1,'']]]
+];