+++ docs/search/pages_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['ailibrary',['AiLibrary',['../index.html',1,'']]]
+];