+++ docs/search/searchdata.js
@@ -1,0 +1,33 @@
+var indexSectionsWithContent =
+{
+  0: "acdefghilmprst",
+  1: "a",
+  2: "aclr",
+  3: "acdefghilmprst",
+  4: "i",
+  5: "adfilmprst",
+  6: "aclt"
+};
+
+var indexSectionNames =
+{
+  0: "all",
+  1: "namespaces",
+  2: "files",
+  3: "functions",
+  4: "defines",
+  5: "groups",
+  6: "pages"
+};
+
+var indexSectionLabels =
+{
+  0: "All",
+  1: "Namespaces",
+  2: "Files",
+  3: "Functions",
+  4: "Macros",
+  5: "Modules",
+  6: "Pages"
+};
+